Moved logout button to Profile
